Delete duplicate items (determined by price)
This commit adds an action that finds items that have been stored in the
database several times. The action then removes every item prior to the
last one.

The way that we determine similar items are:

Offered by:

- The same restaurant
- The same item name
- The same price

If the item is offered again at a new price, we do not delete the old
price.